Author crud

+++ app/Http/Controllers/Author/PostsController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers\Author;
 
 use App\Post;
 use Illuminate\Http\Request;
@@ -15,7 +15,6 @@ use Brian2694\Toastr\Facades\Toastr;
 use App\Category;
 use App\Tag;
 
-
 class PostsController extends Controller
 {
     /**
@@ -25,8 +24,8 @@ class PostsController extends Controller
      */
     public function index()
     {
-        $posts = Post::latest()->get();
-        return view('admin.posts.index', compact('posts'));
+        $posts = Auth::User()->posts()->latest()->get();
+        return view('author.posts.index', compact('posts'));
     }
 
     /**
@@ -35,10 +34,10 @@ class PostsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function create()
-    {   
+    {  
         $categories = Category::all();
         $tags = Tag::all();
-        return view('admin.posts.create', compact('categories', 'tags'));
+        return view('author.posts.create', compact('categories', 'tags'));
     }
 
     /**
@@ -49,7 +48,7 @@ class PostsController extends Controller
      */
     public function store(Request $request)
     {
-       $this->validate($request, [
+         $this->validate($request, [
             'title' => 'required',
             'image' => 'required|mimes:jpeg,bmp,png,jpg',
             'categories' => 'required',
@@ -89,7 +88,7 @@ class PostsController extends Controller
             $post->status = false;
        }
 
-       $post->is_approved = true;
+       $post->is_approved = false;
        $post->save();
 
        //Relationships
@@ -97,7 +96,7 @@ class PostsController extends Controller
        $post->tags()->attach($request->tags);
 
        Toastr::success('Post Saved Successfully :)', 'success');
-       return redirect()->route('admin.post.index');
+       return redirect()->route('author.post.index');
 
     }
 
@@ -108,8 +107,12 @@ class PostsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function show(Post $post)
-    {
-      return view('admin.posts.show', compact('post'));
+    {   
+        if($post->user_id != Auth::id()) {
+          Toastr::error('You are not Authorized to Access this Post', 'Danger');
+          return redirect()->back();
+        }
+        return view('author.posts.show', compact('post'));
     }
 
     /**
@@ -120,9 +123,14 @@ class PostsController extends Controller
      */
     public function edit(Post $post)
     {
+      if($post->user_id != Auth::id()) {
+          Toastr::error('You are not Authorized to Access this Post', 'Danger');
+          return redirect()->back();
+        }
+
         $categories = Category::all();
         $tags = Tag::all();
-        return view('admin.posts.edit', compact('post','categories', 'tags'));
+        return view('author.posts.edit', compact('post','categories', 'tags'));
     }
 
     /**
@@ -134,7 +142,7 @@ class PostsController extends Controller
      */
     public function update(Request $request, Post $post)
     {
-       $this->validate($request, [
+        $this->validate($request, [
             'title' => 'required',
             'image' => 'image',
             'categories' => 'required',
@@ -179,7 +187,7 @@ class PostsController extends Controller
             $post->status = false;
        }
 
-       $post->is_approved = true;
+       $post->is_approved = false;
        $post->save();
 
        //Relationships
@@ -187,28 +195,8 @@ class PostsController extends Controller
        $post->tags()->sync($request->tags);
 
        Toastr::success('Post Updated Successfully :)', 'success');
-       return redirect()->route('admin.post.index');
+       return redirect()->route('author.post.index');
 
-    }
-    
-    public function pending() 
-    {
-      $posts = Post::where('is_approved', false)->get();
-      return view('admin.posts.pending', compact('posts'));
-    }
-
-    public function approve($id)
-    {
-      $post = Post::find($id);
-      if($post->is_approved == false) {
-        $post->is_approved = true;
-        $post->save();
-        Toastr::success('Post Successfully Approved :)', 'success');
-      } else {
-        Toastr::info('Post has already been approved :)', 'Info');
-      }
-
-      return redirect()->back();
     }
 
     /**
@@ -218,10 +206,15 @@ class PostsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy(Post $post)
-    {
+    {   
+        if($post->user_id != Auth::id()) {
+          Toastr::error('You are not Authorized to Access this Post', 'Danger');
+          return redirect()->back();
+        }
+        
         if(Storage::disk('public')->exists('post/'.$post->image))
         {
-        	Storage::disk('public')->delete('post/'.$post->image);
+            Storage::disk('public')->delete('post/'.$post->image);
         }
 
         $post->categories()->detach();
